+++ src/type/observable.ts
@@ -1,12 +1,10 @@
-import type { PlotState } from '@type/state';
-
-export interface Observer {
-  update: (state: PlotState) => void;
+export interface Observer<T> {
+  update: (state: T) => void;
 }
 
-export interface Observable {
-  addObserver: (observer: Observer) => void;
-  removeObserver: (observer: Observer) => void;
-
+export interface Observable<T> {
+  addObserver: (observer: Observer<T>) => void;
+  removeObserver: (observer: Observer<T>) => void;
   notifyStateUpdate: () => void;
+  get state(): T;
 }